+++ h2buster.py
@@ -2,36 +2,68 @@ import hyper
 import threading, queue
 import ssl, sys, time, argparse
 import multiprocessing
+import platform
 
-# Metadata variables, program constants and global variables.
+# Metadata variables
 __author__ = "https://github.com/00xc/"
-__version__ = "0.3b"
+__version__ = "0.3c"
 
 PROGRAM_INFO = "h2buster: an HTTP/2 web directory brute-force scanner."
 DASHLINE = "------------------------------------------------"
 
+# Hardcoded file extensions
+ext = ["/", "" ,".php", ".html", ".asp", ".js", ".css"]
+
+# CLI options metavariable names
 WORDLIST_MVAR = "wordlist"
 TARGET_MVAR = "target"
 DIR_DEPTH_MVAR = "directory_depth"
 CNX_MVAR = "connections"
 THREADS_MVAR = "threads"
+NOCOLOR_MVAR = ""
 
+# CLI options default values.
+# None means required. Boolean means that the option has no argument (either the flag is there or not)
 WORDLIST_DEFAULT = None
 TARGET_DEFAULT = None
 DIR_DEPTH_DEFAULT = 2
 CNX_DEFAULT = 3
 THREADS_DEFAULT = 15
+NOCOLOR_DEFAULT = False # if -nc is present, args.nc = True, otherwise args.nc = False
 
+# CLI options help strings
 WORDLIST_HELP = "Directory wordlist"
 TARGET_HELP = "Target URL/IP address. Default port is 443 and HTTPS enabled. To specify otherwise, use ':port' or 'http://' (port will default to 80 then)."
 DIR_DEPTH_HELP = "Maximum recursive directory depth. Minimum is 1, default is " + str(DIR_DEPTH_DEFAULT) + ", unlimited is 0."
 CNX_HELP = "Number of HTTP/2 connections. Default is " + str(CNX_DEFAULT) + "."
 THREADS_HELP = "Number of threads per connection. Default is " + str(THREADS_DEFAULT) + "."
+NOCOLOR_HELP = "Disable colored output text."
 
+# Other hardcoded values
 TIMESTAMP_ROUND = 3
 TLS_DEFAULT = 1
 
-ext = ["/", "" ,".php", ".html", ".asp", ".js", ".css"]
+# Define colors (only Linux)
+if platform.system() == "Linux" or platform.system() == "Darwin":
+	COLOR_200 = '\033[92m' # green
+	COLOR_302 = '\033[94m' # blue
+	COLOR_301 = '\033[94m' # blue
+	COLOR_403 = '\033[93m' # yellow
+	COLOR_ERROR = '\033[91m' # red
+	COLOR_END = '\033[0m' # end color
+
+# Function: returns colored string according to status
+def colorstring(s, status=0):
+	global NOCOLOR
+	if status == 0 or NOCOLOR==True: return s
+	if platform.system() == "Linux":
+		try:
+			start = globals()["COLOR_" + str(status)]
+			end = COLOR_END
+			return "".join([start, s, end])
+		except KeyError:
+			return s
+	else: return s
 
 # Function: return time since t0
 def timestamp(t0):
@@ -41,7 +73,6 @@ def timestamp(t0):
 # Function: return True if entry is a directory, False otherwise.
 def isdir(e):
 	if e == "/": return False
-
 	if len(e)>0 and e[-1]=="/":
 		if e.count(".") > 0:
 			if any([len(x)==0 for x in e.split(".")]):
@@ -52,13 +83,18 @@ def isdir(e):
 	else:
 		return False
 
-# Function: read "opts" options from command line interface
+# Function: read "opts" options from CLI. 
 def read_inputs(info, opts, h, defaults, mvar):
 	parser = argparse.ArgumentParser(description=info)
-	for i, o in enumerate(opts):
-		if defaults[i]==None: req = True
-		else: req = False
-		parser.add_argument("-"+o, help=h[i], default=defaults[i], required=req, metavar=mvar[i])
+	for default, o, htext, mtext in zip(defaults, opts, h, mvar):
+		if isinstance(default, bool):
+			if default == True: action = "store_false"
+			else: action = "store_true"
+			parser.add_argument("-"+o, help=htext, default=default, action=action)
+		else:
+			if default==None: req = True
+			else: req = False
+			parser.add_argument("-"+o, help=htext, default=default, required=req, metavar=mtext)
 	args = parser.parse_args()
 	return args
 
@@ -72,7 +108,7 @@ def parse_target(target):
 		url = target[1]
 		if target[0] == "http": s = 0
 		elif target[0] == "https": s = 1
-		else: sys.exit("[-] Target not understood")
+		else: sys.exit(colorstring("[-] Target not understood", status="ERROR"))
 
 	url = url.split("/", 1)
 	ip = url[0]
@@ -92,7 +128,7 @@ def h2_connect(s, ip):
 		elif s==0: port=80
 	else:
 		try: port = int(ip.split(":")[-1])
-		except ValueError: sys.exit("[-] Invalid URL")
+		except ValueError: sys.exit(colorstring("[-] Invalid URL", status="ERROR"))
 	# Start connection
 	if s == 1:
 		ctx = ssl.SSLContext()
@@ -102,7 +138,9 @@ def h2_connect(s, ip):
 	elif s == 0:
 		conn = hyper.HTTP20Connection(ip, port=port, enable_push=False)
 	try: conn.connect()
-	except AssertionError: sys.exit("[-] H2 not supported for that target.")
+	except AssertionError:
+		conn.close()
+		sys.exit(colorstring("[-] H2 not supported for that target.", status="ERROR"))
 	return conn, port
 
 # Function: main scan function. Starts up a number of processes which handle their own h2 connection and sends them entries to scan
@@ -112,7 +150,7 @@ def main_scan(s, ip, directory, wordlist, dir_depth, max_depth, connections, thr
 		if dir_depth >= max_depth: return
 	global ext
 
-	print("\n[*] Starting scan on " + directory)
+	print(colorstring("\n[*] Starting scan on " + directory))
 
 	# Start input/output queues
 	manager = multiprocessing.Manager()
@@ -140,7 +178,7 @@ def main_scan(s, ip, directory, wordlist, dir_depth, max_depth, connections, thr
 		for i in range(connections*threads):
 			inwork.put((None, None))
 		for p in pool: p.join()
-		sys.exit("[-] Wordlist file not found.")
+		sys.exit(colorstring("[-] Wordlist file not found.", status="ERROR"))
 
 	# Send kill signals
 	for i in range(connections*threads):
@@ -185,7 +223,7 @@ def thread_worker(conn, inwork, output):
 				tail = ""
 				if st==200 and isdir(entry):
 					output.append(directory + entry)
-			print("".join([directory, entry, ": ", str(st), tail]))
+			print(colorstring("".join([directory, entry, ": ", str(st), tail]), status=st))
 
 # Main start point. Read, verify inputs and call main_scan()
 if __name__ == '__main__':
@@ -195,11 +233,15 @@ if __name__ == '__main__':
 	print(DASHLINE)
 
 	# Read CLI inputs
-	opts = ["w", "u", "r", "c", "t"]
-	mvar = [WORDLIST_MVAR, TARGET_MVAR, DIR_DEPTH_MVAR, CNX_MVAR, THREADS_MVAR]
-	h = [WORDLIST_HELP, TARGET_HELP, DIR_DEPTH_HELP, CNX_HELP, THREADS_HELP]
-	defaults = [WORDLIST_DEFAULT, TARGET_DEFAULT, DIR_DEPTH_DEFAULT, CNX_DEFAULT, THREADS_DEFAULT]
+	opts = ["w", "u", "r", "c", "t", "nc"]
+	mvar = [WORDLIST_MVAR, TARGET_MVAR, DIR_DEPTH_MVAR, CNX_MVAR, THREADS_MVAR, NOCOLOR_MVAR]
+	h = [WORDLIST_HELP, TARGET_HELP, DIR_DEPTH_HELP, CNX_HELP, THREADS_HELP, NOCOLOR_HELP]
+	defaults = [WORDLIST_DEFAULT, TARGET_DEFAULT, DIR_DEPTH_DEFAULT, CNX_DEFAULT, THREADS_DEFAULT, NOCOLOR_DEFAULT]
 	args = read_inputs(PROGRAM_INFO, opts, h, defaults, mvar)
+	
+	# Set nc as global constant so we don't have to pass it around for every function that prints text.
+	# It's hacky, but whatever.
+	NOCOLOR = args.nc
 
 	# Input checking
 	try:
@@ -207,15 +249,26 @@ if __name__ == '__main__':
 		args.c = int(args.c)
 		args.t = int(args.t)
 		if args.t<1 or args.c<1 or args.r<0:
-			sys.exit("[-] " + CNX_MVAR + " and " + THREADS_MVAR + " must be greater than zero. " + DIR_DEPTH_MVAR + " must be greater than or equal to zero.")
+			sys.exit(colorstring("[-] " + CNX_MVAR + " and " + THREADS_MVAR + " must be greater than zero. " + DIR_DEPTH_MVAR + " must be greater than or equal to zero.", status="ERROR"))
 	except ValueError:
-		sys.exit("[-] Invalid non-numerical option introduced")
+		sys.exit(colorstring("[-] Invalid non-numerical option introduced", status="ERROR"))
 
 	# Start timer (for benchmarking purposes)
 	t0 = time.time()
 
 	# Parse target URL
 	ip, start_dir, s = parse_target(args.u)
+
+	# Check if target accepts requests and supports H2.
+	conn, port = h2_connect(s, ip)
+	try:
+		sid = conn.request("HEAD", "/")
+		resp = conn.get_response(sid)
+	except ConnectionResetError:
+		sys.exit(colorstring("[-] Connection reset. Are you sure target supports HTTP/2?", status="ERROR"))
+	finally:
+		conn.close()
+	print(colorstring("[+] Target supports HTTP/2", status=200))
 
 	# Print info
 	print("[*] Initializing scan on " + ip)